+++ env/bootstrap.sh
@@ -165,6 +165,20 @@ apt-get -y install xvfb
 apt-get -y install firefox
 apt-get -y install memcached
 
+# install firefox45 and geckodriver for selenium
+# firefox45
+wget https://ftp.mozilla.org/pub/firefox/releases/45.0/linux-x86_64/en-US/firefox-45.0.tar.bz2
+tar -xjf firefox-45.0.tar.bz2
+sudo rm -rf  /opt/firefox
+sudo mv firefox /opt/firefox45
+sudo mv /usr/bin/firefox /usr/bin/firefoxold
+sudo ln -s /opt/firefox45/firefox /usr/bin/firefox
+# geckodriver
+wget https://github.com/mozilla/geckodriver/releases/download/v0.13.0/geckodriver-v0.13.0-linux64.tar.gz
+tar -xvzf geckodriver*
+chmod +x geckodriver
+sudo mv geckodriver /usr/local/bin/
+
 # download and hack broken zlib
 apt-get -y install zlib1g-dev
 # cd /lib